Add files via upload
Hangman is a classic word-guessing game where players try to guess a hidden word by suggesting letters one at a time. Each incorrect guess brings a step closer to completing a drawing of a hangman. The goal is to guess the word before the drawing is completed. It's a fun way to test vocabulary and spelling skills!